7/11/2021 > Added age and sex to patient's information > Added option to select vaccinator for the second dosage > Removed city column from Patient Records Table, and added the name of the second vaccinator > Fixed Update and Delete functionalities in Account Management Table > Vaccine's StorageAmount now decrements/increments when updating/creating new patient records > Vaccine deletion would not work if there are still patients using that vaccine (TODO for next commit: patients with complete status should not be counted as patients that still use the vaccine)

+++ src/vaccine/frontend/controllers/addVaccine.java
@@ -78,7 +78,15 @@ public class addVaccine implements Initializable {
             Optional<ButtonType> action = alert.showAndWait();
 
             if (action.get() == ButtonType.OK) {
-                vaccineDAO.deleteVaccine(vaccine);
+                if (scheduleDAO.getPatientCountBasedOnVaccineID(vaccine.getVaccineID()) == 0){
+                    vaccineDAO.deleteVaccine(vaccine);
+                } else {
+                    alert = new Alert(Alert.AlertType.ERROR);
+                    alert.setContentText("Error! You cannot delete vaccine records while Patient Records that uses the vaccine exists!");
+                    alert.setTitle("Error");
+                    alert.show();
+                }
+
                 Stage stage = (Stage) main.getScene().getWindow();
                 stage.hide();
